+++ ChangeLog.md
@@ -1,4 +1,6 @@
 ## 0.1.0.1
+### Added
+- New "resizing-cells" example to show what happens when cells resize. This will be useful to demonstrate that virtual grids work sensibly.
 ### Fixed
 - The footer now only takes up vertical space if there are some footer widgets defined.